Expose our schema on /meta/schema URL
We are gravitating towards "self-describing" services, which we can
interrogate for metadata

One of the metadata is a "data dictionary"; we can start from a
generated schema to enable generating a data dictionary.

This change exposes our database schema on `/meta/schema`.

Things to note:

- generates the schema and the entity-relationship diagram at _runtime_
- `/meta/schema` is configured to allow unauthenticated access
- caches the first rendering in a temporary file until restart
- toggled by a feature called called `FEATURE_SCHEMAINFO`

Trade-offs with running this at runtime vs build-time:

- runtime: was quicker to build
- runtime: but requires `graphviz` in the final image