Change the circleCI config to release from the `release/code-freeze` branch instead of main
The reason for doing this is that we want to revert the release to a previous known
safe version of the code. We also want to be unblocked from releasing small critical
bug fixes or CVE updates without being blocked by queued changes in `main`. The
`release/code-freeze` branch was taken from the previous release, with some critical
library upgrades cherry-picked.