Time delay between referrals for same person/service
A key business metric we are trying to improve is "waste", characterised
by probation staff having to re-refer someone to the same service.

Our goal as a team is to make sure the "delays" between referrals are
kept long -- which indicate actual needs; and eliminate the need to
"quickly" re-refer.

This metric will be published to Prometheus and AppInsights since both
are wired into micrometer.