⚗️  Track follow-up referrals for same person
Introduce tracking "follow-up" referrals along "repeat" referrals.

Definitions:

- "repeated" referral: new referral to same person and intervention
- "redirected" referral: new referral to same person

The metric deliberately does **not** track the intervention IDs: at
first, we want to track aggregate counts, not transition patterns.

We might choose to not increase the metric if the previous referral
is too old to indicate re-referrals (once we see the data)

---

We received user research feedback on the previous metric:

> yes, but just note that sometimes re-referrals happen because they
> have selected the wrong intervention type…e.g. Social inclusion for
> those who are not in custody
>
> that is a big one
>
> that will be one of the bigger reasons for a re-referral in the first
> 7 days i think

This metric implements the above need (measuring re-referrals on
different interventions)