Expose 5xx calls on services we directly depend on
So that we can be aware of dependencies having problems even when it's
not through the service. Superagent in UI does not (yet?) expose
metrics about client calls.

This does not include community-api, as that is not running in the
Cloud Platform.

Prometheus query: `max(rate(nginx_ingress_controller_requests{status=~"5..",exported_namespace=~"(token-verification-api-prod|hmpps-auth-prod|hmpps-assess-risks-and-needs-prod)"}[2m])*120) by(exported_service)`