Dereference end-of-service-report at publishing
Often enough, our `ApplicationListener` classes blow up with a
`NullPointerException` due to some state assumption.

We need to detect assumptions about the state earlier. The likely best
place to do so is at publishing because we know _exactly_ what state we
expect when we, e.g. fire an event about a submitted end-of-service
report.

This change uses `EndOfServiceReportEvent` as the example, as it only
has one type (submitted).

The related event, publisher, and listener classes have all been pulled
into standalone classes with matching test names.