Make frequency of requests visible on heatmaps
Changes heatmap from a single-colour scheme with a square root scale
(good for showing if anything is happening) to a spectral colour scheme
(good for identifying is something happens often or not).

Also hides zeroes -- that will remove the "background" image and only
render colour if there is a value.