Create "delivery state" abstraction
So that a lot of the repeated complexity becomes clearer in code.

We keep talking about "when delivery..." when describing the problem,
but the code was not showing that.

Having an abstraction separating "what delivery state we are in", "do we
need end of service report" and "what state will that transition to"
makes it easier to read/reason about this code.

No behaviour changes; however, we _really_ should be using
`approvedActionPlan` instead of `currentActionPlan`