Improve flaky retrying client test
On CircleCI, this test sometimes fails.

The first canned response, which should be 500, instead comes back with
a read timeout:

```
15:40:53.853 [Test worker] DEBUG org.apache.http.wire - http-outgoing-0 >> "grant_type=client_credentials"
15:40:53.909 [Test worker] DEBUG org.apache.http.wire - http-outgoing-0 << "[read] I/O error: Read timed out"
```

In successful cases, it looks like this:

```
14:24:24.889 [Test worker] DEBUG org.apache.http.wire - http-outgoing-0 >> "grant_type=client_credentials"
14:24:24.904 [Test worker] DEBUG org.apache.http.wire - http-outgoing-0 << "HTTP/1.1 500 Server Error[\r][\n]"
```

Increasing the read timeout to 100ms should allow CircleCI to be _less_
flaky. Unfortunately, longer timeouts cause other issues. So let's try
it this way and see if it improves.