Add support for hotfix branches
This change enables developers to deploy hotfixes to production without accidentally deploying incomplete changes from the main branch. Ideally, the main branch would never be blocked, however there are times when changes have been merged that aren't ready for production - this PR offers an escape hatch for when things break in production and need to be fixed quickly.

To create a hotfix,
1. Checkout the commit id that is currently deployed to production
2. Create a branch with the prefix `hotfix/`, e.g. `hotfix/put-out-fire`
3. Create a pull request into the `main` branch

The code changes on the hotfix branch will then go through the pipeline as normal and can be approved into production.

Note: this will not disrupt your usual workflow, `hotfix/*` branches only need to be used in exceptional circumstances (i.e. when there are changes in the main branch that aren't ready to be deployed, but you urgently need to fix something in production).