Update data extractor to latest (#2108)
We had a critical alert from dependabot in the data extractor repository which required us to bump a few dependencies and create a new release. This PR bumps the data extractor to the latest release. We've tested this with the calculate-release-dates service and it worked without issue. Hope this is ok - happy to discuss. Thanks!

Co-authored-by: guyvw1 <89020559+guyvw1@users.noreply.github.com>